LPS-88869 Move to its own package
