+++ portal-web/docroot/html/portal/init.jsp
@@ -36,10 +36,10 @@ page import="com.liferay.portal.kernel.parsers.bbcode.BBCodeTranslatorUtil" %><%
 page import="com.liferay.portal.kernel.portlet.PortletConfigurationLayoutUtil" %><%@
 page import="com.liferay.portal.kernel.servlet.HttpHeaders" %><%@
 page import="com.liferay.portal.kernel.templateparser.TransformException" %><%@
+page import="com.liferay.portal.kernel.term.TermsOfUseContentProvider" %><%@
+page import="com.liferay.portal.kernel.term.TermsOfUseContentProviderUtil" %><%@
 page import="com.liferay.portal.kernel.util.ClassUtil" %><%@
 page import="com.liferay.portal.kernel.util.ProgressTracker" %><%@
-page import="com.liferay.portal.kernel.util.TermsOfUseContentProvider" %><%@
-page import="com.liferay.portal.kernel.util.TermsOfUseContentProviderUtil" %><%@
 page import="com.liferay.portal.setup.SetupWizardUtil" %><%@
 page import="com.liferay.portal.util.LicenseUtil" %><%@
 page import="com.liferay.portal.util.PortletCategoryUtil" %><%@
